Feat: 기능에 따른 친구 / 익명 피드 분리
