+++ navigations/AppTabNavigator/PostTab/MainPostTab.js
@@ -1,13 +1,9 @@
 import React from 'react';
-import { View } from 'react-native';
 import { createBottomTabNavigator } from "@react-navigation/bottom-tabs";
 import { FriendPosts, PublicPosts } from './index';
-import FeedButton from "../../../components/FeedButton";
 
 const MainPostTab = createBottomTabNavigator();
 export const MainPostTabScreen = () => {
-	const friendPostsRef = React.useRef(); // 탭의 활성화 확인
-	const PublicPostsRef = React.useRef(); // 탭의 활성화 확인
 	
 	return (
 		<MainPostTab.Navigator
@@ -18,86 +14,13 @@ export const MainPostTabScreen = () => {
 		>
 			<MainPostTab.Screen
 				name="FriendFeed"
-				children={(props) => <FriendPosts ref={friendPostsRef} {...props} />}
-				options={({ navigation }) => ({
-					headerTitle: () => {
-						// 현재 활성화된 탭 이름 가져오기
-						const currentRouteName = navigation.getState().routes[navigation.getState().index].name;
-						return (
-							<View style={{ flexDirection: 'row' }}>
-								<FeedButton
-									title="친구"
-									onPress={() => {
-										if (currentRouteName === 'FriendFeed') { // 현재 활성화된 탭과 비교
-											friendPostsRef.current.getPosts(); // FriendFeed의 새로고침 함수 호출
-										} else {
-											navigation.navigate('FriendFeed'); // 아니면 이동
-										}
-									}}
-									isSelected={currentRouteName === 'FriendFeed'} // FeedButton 스타일 전환
-								/>
-								<FeedButton
-									title="더보기"
-									onPress={() => {
-										if (currentRouteName === 'OtherFeed') { // 현재 활성화된 탭과 비교
-											PublicPostsRef.current.getPosts(); // FriendFeed의 새로고침 함수 호출
-										} else {
-											navigation.navigate('OtherFeed'); // 아니면 이동
-										}
-									}}
-									isSelected={currentRouteName === 'OtherFeed'} // FeedButton 스타일 전환
-								/>
-							</View>
-						);
-					},
-					headerTitleAlign: "center",
-					headerStyle: {
-						height: 30,
-					}
-				})}
+				component={FriendPosts}
 			>
 			</MainPostTab.Screen>
 			<MainPostTab.Screen
 				name="OtherFeed"
-				children={(props) => <PublicPosts ref={PublicPostsRef} {...props} />}
-				options={({ navigation }) => ({
-					headerTitle: () => {
-						// 현재 활성화된 탭 이름 가져오기
-						const currentRouteName = navigation.getState().routes[navigation.getState().index].name;
-						return (
-							<View style={{ flexDirection: 'row' }}>
-								<FeedButton
-									title="친구"
-									onPress={() => {
-										if (currentRouteName === 'FriendFeed') { // 현재 활성화된 탭과 비교
-											PublicPostsRef.current.getPosts(); // FriendFeed의 새로고침 함수 호출
-										} else {
-											navigation.navigate('FriendFeed'); // 아니면 이동
-										}
-									}}
-									isSelected={currentRouteName === 'FriendFeed'} // FeedButton 스타일 전환
-								/>
-								<FeedButton
-									title="더보기"
-									onPress={() => {
-										if (currentRouteName === 'OtherFeed') { // 현재 활성화된 탭과 비교
-											PublicPostsRef.current.getPosts(); // FriendFeed의 새로고침 함수 호출
-										} else {
-											navigation.navigate('OtherFeed'); // 아니면 이동
-										}
-									}}
-									isSelected={currentRouteName === 'OtherFeed'} // FeedButton 스타일 전환
-								/>
-							</View>
-						);
-					},
-					headerTitleAlign: "center",
-					headerStyle: {
-						height: 30,
-					}
-				})}
-			>
-			</MainPostTab.Screen>
+				component={PublicPosts}
+			/>
 		</MainPostTab.Navigator>
 	);
 }
